+++ CHANGES.md
@@ -1,3 +1,7 @@
+
+- Add `Line_break`, `Grapheme_cluster_break`, `Word_break` and
+  `Sentence_break` properties.
+
 v0.9.0 2014-06-28 Cambridge (UK)
 -------------------------------